chore: add config to vega wallet provider
