+++ libs/wallet/src/connect-dialog/chrome-icon.tsx
@@ -1,0 +1,77 @@
+export const ChromeIcon = () => {
+  return (
+    <svg
+      width="28"
+      height="28"
+      viewBox="0 0 32 32"
+      fill="none"
+      data-testid="chrome-logo"
+    >
+      <g clipPath="url(#clip0_3681_24659)">
+        <path
+          d="M15.9987 9.99963L26.3893 9.99964C25.3364 8.17534 23.8217 6.6604 21.9976 5.60716C20.1735 4.55391 18.1042 3.99949 15.9979 3.99964C13.8915 3.99979 11.8223 4.5545 9.99837 5.608C8.1744 6.66149 6.65995 8.17664 5.6073 10.0011L10.8026 18.9996L10.8072 18.9984C10.2787 18.0871 9.99984 17.0525 9.99865 15.999C9.99747 14.9454 10.274 13.9102 10.8005 12.9977C11.3269 12.0851 12.0847 11.3275 12.9973 10.8011C13.9099 10.2748 14.9451 9.99832 15.9987 9.99963Z"
+          fill="url(#paint0_linear_3681_24659)"
+        />
+        <path
+          d="M21.1974 18.9989L16.0021 27.9974C18.1084 27.9977 20.1777 27.4435 22.0019 26.3904C23.8261 25.3373 25.3409 23.8224 26.3939 21.9982C27.447 20.174 28.0012 18.1047 28.0008 15.9983C28.0004 13.892 27.4455 11.8228 26.3918 9.99898L16.0012 9.99899L15.9999 10.0036C17.0534 10.0016 18.0889 10.2773 19.0018 10.8031C19.9148 11.3288 20.673 12.0859 21.2001 12.9981C21.7272 13.9103 22.0044 14.9454 22.004 15.9989C22.0035 17.0524 21.7253 18.0872 21.1974 18.9989Z"
+          fill="url(#paint1_linear_3681_24659)"
+        />
+        <path
+          d="M10.8044 19.0016L5.60914 10.0031C4.5557 11.8271 4.00106 13.8963 4.00098 16.0026C4.0009 18.109 4.55539 20.1782 5.60869 22.0023C6.66199 23.8264 8.17698 25.341 10.0013 26.3938C11.8257 27.4467 13.895 28.0007 16.0014 28.0001L21.1967 19.0015L21.1933 18.9981C20.6683 19.9115 19.9118 20.6703 19 21.1981C18.0882 21.7259 17.0534 22.004 15.9999 22.0043C14.9464 22.0047 13.9114 21.7273 12.9992 21.2001C12.0871 20.6729 11.3301 19.9146 10.8044 19.0016Z"
+          fill="url(#paint2_linear_3681_24659)"
+        />
+        <path
+          d="M16 22C19.3137 22 22 19.3137 22 16C22 12.6863 19.3137 10 16 10C12.6863 10 10 12.6863 10 16C10 19.3137 12.6863 22 16 22Z"
+          fill="white"
+        />
+        <path
+          d="M16 20.75C18.6234 20.75 20.75 18.6234 20.75 16C20.75 13.3766 18.6234 11.25 16 11.25C13.3766 11.25 11.25 13.3766 11.25 16C11.25 18.6234 13.3766 20.75 16 20.75Z"
+          fill="#1A73E8"
+        />
+      </g>
+      <defs>
+        <linearGradient
+          id="paint0_linear_3681_24659"
+          x1="25.093"
+          y1="9.25084"
+          x2="14.702"
+          y2="27.2485"
+          gradientUnits="userSpaceOnUse"
+        >
+          <stop stopColor="#D93025" />
+          <stop offset="1" stopColor="#EA4335" />
+        </linearGradient>
+        <linearGradient
+          id="paint1_linear_3681_24659"
+          x1="27.073"
+          y1="11.4997"
+          x2="6.29104"
+          y2="11.4997"
+          gradientUnits="userSpaceOnUse"
+        >
+          <stop stopColor="#FCC934" />
+          <stop offset="1" stopColor="#FBBC04" />
+        </linearGradient>
+        <linearGradient
+          id="paint2_linear_3681_24659"
+          x1="17.2992"
+          y1="27.2508"
+          x2="6.90819"
+          y2="9.25305"
+          gradientUnits="userSpaceOnUse"
+        >
+          <stop stopColor="#1E8E3E" />
+          <stop offset="1" stopColor="#34A853" />
+        </linearGradient>
+        <clipPath id="clip0_3681_24659">
+          <rect
+            width="24"
+            height="24"
+            fill="white"
+            transform="translate(4 4)"
+          />
+        </clipPath>
+      </defs>
+    </svg>
+  );
+};
